+++ tencentcloud/antiddos/v20200309/models.go
@@ -1843,7 +1843,7 @@ type DDoSAIRelation struct {
 type DDoSGeoIPBlockConfig struct {
 
 	// 区域类型，取值[
-	// oversea(海外)
+	// oversea(境外)
 	// china(国内)
 	// customized(自定义地区)
 	// ]