+++ tencentcloud/tke/v20180525/errors.go
@@ -197,6 +197,12 @@ const (
 	// 任务未找到。
 	INTERNALERROR_TASKNOTFOUND = "InternalError.TaskNotFound"
 
+	// 询价错误。
+	INTERNALERROR_TRADECOMMON = "InternalError.TradeCommon"
+
+	// 账户余额不足。
+	INTERNALERROR_TRADEINSUFFICIENTBALANCE = "InternalError.TradeInsufficientBalance"
+
 	// 内部错误。
 	INTERNALERROR_UNEXCEPTEDINTERNAL = "InternalError.UnexceptedInternal"